worked out SubjectJob
