+++ eegprep/jobs/concat_epochs.py
@@ -1,0 +1,7 @@
+from eegprep.jobs.base import BaseJob
+
+
+class ConcatEpochsJob(BaseJob):
+    
+    def run(self):
+        pass
